Using corporate GitHub identity as author for GitHub-hosted projects.
The individual who performed the commit is used as the committer.

This may be an inversion of the expected use where author is the person
who made the changes and committer is the person who checked-off on the
commit.  However, we are chasing maximum GitHub clout, and want all our
contributions to be under our corporate identity.  We also believe this
to be consistent with a corporation being a legal entity that can exist
outside of and outlive the individuals who comprise it at any given
moment.

This is achieved through setting GIT_AUTHOR_NAME and GIT_AUTHOR_EMAIL
environment variables via .envrc files using the "direnv" utility.